fixing error at cart

+++ controllers/cart_controller.go
@@ -12,13 +12,13 @@ import (
 )
 
 func GetAuthenticatedUserID(c *gin.Context) (int64, error) {
-	UserID, exist := c.Get("UserID")
+	userID, exist := c.Get("user_id")
 	if !exist {
 		return 0, gorm.ErrRecordNotFound
 	}
 
 	// Convert value to int64
-	id, ok := UserID.(int64)
+	id, ok := userID.(int64)
 	if !ok {
 		return 0, gorm.ErrInvalidData
 	}
@@ -30,7 +30,7 @@ func AddItemToCart(db *gorm.DB) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		// Bind request body JSON to struct
 		var request models.CartRequest
-		err := c.ShouldBindBodyWithJSON(&request)
+		err := c.ShouldBindJSON(&request)
 		if err != nil {
 			responses.Error(c, http.StatusBadRequest, "Invalid request body", err.Error())
 			return
@@ -51,7 +51,7 @@ func AddItemToCart(db *gorm.DB) gin.HandlerFunc {
 		var cart models.Cart
 		if err := db.Where("user_id = ?", userID).First(&cart).Error; err != nil {
 			if err == gorm.ErrRecordNotFound {
-				cart := models.Cart{UserID: userID}
+				cart = models.Cart{UserID: userID}
 				if err := db.Create(&cart).Error; err != nil {
 					responses.Error(c, http.StatusInternalServerError, "Failed to create cart", err.Error())
 					return
@@ -107,7 +107,8 @@ func AddItemToCart(db *gorm.DB) gin.HandlerFunc {
 			log.Printf("Warning: Could not preload for cart item %d: %v", cartItem.ID, err)
 		}
 		responses.Success(c, http.StatusOK, "Add item to cart success!", gin.H{
-			"message": "Add item to cart successfuly",
+			"message":  "Add item to cart successfuly!",
+			"cartItem": cartItem,
 		})
 	}
 }
@@ -120,6 +121,7 @@ func GetCart(db *gorm.DB) gin.HandlerFunc {
 			responses.Error(c, http.StatusUnauthorized, "Unauthorized", "User not authenticated")
 			return
 		}
+
 		var cart models.Cart
 		if err := db.Preload("CartItems.Product").Where("user_id = ?", userID).First(&cart).Error; err != nil {
 			if err == gorm.ErrRecordNotFound {
@@ -146,7 +148,7 @@ func UpdateCartItemQuantity(db *gorm.DB) gin.HandlerFunc {
 		// Bind request body JSON to struct
 		var request models.CartRequest
 
-		if err := c.ShouldBindBodyWithJSON(&request); err != nil {
+		if err := c.ShouldBindJSON(&request); err != nil {
 			responses.Error(c, http.StatusBadRequest, "Invalid request", err.Error())
 			return
 		}
@@ -161,11 +163,11 @@ func UpdateCartItemQuantity(db *gorm.DB) gin.HandlerFunc {
 			responses.Error(c, http.StatusUnauthorized, "Unauthorized", "User not authenticated")
 			return
 		}
+
 		// Make sure cart item belong to correct user & correct shopping cart
 		var cartItem models.CartItem
 
-		err = db.Joins("JOIN cart ON cart.id = cart_items.cart_id").
-			Where("cart_item = ? AND carts.user_id = ?", cartItemID, userID).First(&cartItem).Error
+		err = db.Preload("Cart").First(&cartItem, cartItemID).Error
 
 		if err != nil {
 			if err == gorm.ErrRecordNotFound {
@@ -176,24 +178,33 @@ func UpdateCartItemQuantity(db *gorm.DB) gin.HandlerFunc {
 			return
 		}
 
+		// Verification cart to authenticated user
+		if cartItem.Cart.UserID != int64(userID) {
+			responses.Error(c, http.StatusUnauthorized, "Unauthorized", "Cart is not yours")
+			return
+		}
+
 		if request.Quantity == 0 {
 			if err := db.Delete(&cartItem).Error; err != nil {
 				responses.Error(c, http.StatusInternalServerError, "Failed to remove item from cart", err.Error())
 				return
 			}
-			responses.Error(c, http.StatusNoContent, "Status no content", "Status no content")
+			responses.Success(c, http.StatusNoContent, "Item Remove Sucessfully", nil)
 			return
 		}
+
 		cartItem.Quantity = int(request.Quantity)
 		if err := db.Save(&cartItem).Error; err != nil {
 			responses.Error(c, http.StatusInternalServerError, "Failed to update cart item", err.Error())
 			return
 		}
+
 		if err := db.Preload("Product").First(&cartItem, cartItemID).Error; err != nil {
 			log.Printf("Warning: Could not preload product for cart item %d: %v", cartItem.ID, err)
 		}
 		responses.Success(c, http.StatusOK, "Update cart item successfuly", gin.H{
-			"message": "Update cart item successfuly",
+			"message":  "Update cart item successfuly",
+			"cartItem": cartItem,
 		})
 	}
 }
@@ -216,9 +227,7 @@ func RemoveCartItem(db *gorm.DB) gin.HandlerFunc {
 		// Make sure cart item is belong to correct owner & correct shopping cart
 		var cartItem models.CartItem
 
-		err = db.Joins("JOIN carts on carts.id = cart_items.cart_id").
-			Where("cart_items.id = ? AND carts.user_id = ?", cartItemID, userID).
-			First(&cartItem).Error
+		err = db.Preload("Cart").First(&cartItem, cartItemID).Error
 
 		if err != nil {
 			if err == gorm.ErrRecordNotFound {
@@ -229,11 +238,16 @@ func RemoveCartItem(db *gorm.DB) gin.HandlerFunc {
 			return
 		}
 
+		// Verfication cart to authenticated user
+		if cartItem.Cart.UserID != int64(userID) {
+			responses.Error(c, http.StatusUnauthorized, "Unauthorized", "This cart is not yours")
+		}
+
 		if err := db.Delete(&cartItem).Error; err != nil {
 			responses.Error(c, http.StatusInternalServerError, "Failed to delete cart item", err.Error())
 			return
 		}
 
-		c.Status(http.StatusNoContent) // 204 no content for success deleting cart item
+		responses.Success(c, http.StatusNoContent, "Item remove sucessfully!", nil)
 	}
 }
